Move KerasAudioClassifer using 1D CNN to experimental
